Convert to typescript-react-router-5 plygin. Remove GeneratorReactRouterV5
