--- packages/core/src/generate/generateAppFiles/generatorReactRouterV5/generateLinkFileReactRouterV5.test.ts
@@ -1,112 +1,0 @@
-import { generateLinkFileReactRouterV5, GenerateLinkFileReactRouterV5Params } from "./generateLinkFileReactRouterV5";
-
-describe("generateLinkFileReactRouterV5", () => {
-  const defaultParams: GenerateLinkFileReactRouterV5Params = {
-    importGenerateUrl: { namedImports: [{ name: "generateUrl" }], from: "route-codegen" },
-    routeLinkOption: {
-      importLink: {
-        from: "src/common/Link",
-        defaultImport: "Link",
-        namedImports: [{ name: "CustomLinkProps" }],
-      },
-      linkComponent: "Link",
-      linkProps: "CustomLinkProps",
-      hrefProp: "to",
-      generateLinkComponent: true,
-      generateRedirectComponent: true,
-      generateUseRedirect: true,
-      generateUseParams: true,
-      mode: "strict",
-    },
-    routeName: "Login",
-    patternNamedExports: {
-      originName: "originLogin",
-      filename: "patternLogin",
-      patternName: "patternLogin",
-      urlParamsInterfaceName: "UrlParamsLogin",
-      patternNameNextJS: "patternNextJSLogin",
-    },
-    destinationDir: "path/to/routes",
-  };
-
-  describe("ReactRouterV5", () => {
-    it("should generate correctly if no path params", () => {
-      const templateFile = generateLinkFileReactRouterV5({ ...defaultParams });
-
-      expect(templateFile.filename).toBe("LinkLogin");
-      expect(templateFile.extension).toBe(".tsx");
-      expect(templateFile.destinationDir).toBe("path/to/routes");
-      expect(templateFile.template).toMatchInlineSnapshot(`
-        "import React from 'react'
-          import {generateUrl,} from 'route-codegen'
-          import Link, {CustomLinkProps,} from 'src/common/Link'
-          import {patternLogin,UrlParamsLogin,} from './patternLogin'
-          type LinkLoginProps = Omit<CustomLinkProps, 'to'> & { urlParams?: UrlParamsLogin }
-          export const LinkLogin: React.FunctionComponent<LinkLoginProps> = ({ urlParams, ...props }) => {
-            const to = generateUrl(patternLogin, { path: {}, query: urlParams?.query, origin: urlParams?.origin });
-            return <Link {...props} to={to} />;
-          }"
-      `);
-    });
-
-    it("should generate correctly with path params", () => {
-      const templateFile = generateLinkFileReactRouterV5({
-        ...defaultParams,
-        patternNamedExports: {
-          ...defaultParams.patternNamedExports,
-          pathParamsInterfaceName: "PathParamsLogin",
-        },
-      });
-
-      expect(templateFile.filename).toBe("LinkLogin");
-      expect(templateFile.extension).toBe(".tsx");
-      expect(templateFile.destinationDir).toBe("path/to/routes");
-      expect(templateFile.template).toMatchInlineSnapshot(`
-        "import React from 'react'
-          import {generateUrl,} from 'route-codegen'
-          import Link, {CustomLinkProps,} from 'src/common/Link'
-          import {patternLogin,UrlParamsLogin,} from './patternLogin'
-          type LinkLoginProps = Omit<CustomLinkProps, 'to'> & { urlParams: UrlParamsLogin }
-          export const LinkLogin: React.FunctionComponent<LinkLoginProps> = ({ urlParams, ...props }) => {
-            const to = generateUrl(patternLogin, { path: urlParams.path, query: urlParams?.query, origin: urlParams?.origin });
-            return <Link {...props} to={to} />;
-          }"
-      `);
-    });
-
-    it("should generate correctly with named component import", () => {
-      const templateFile = generateLinkFileReactRouterV5({
-        ...defaultParams,
-        routeLinkOption: {
-          importLink: {
-            from: "src/common/Link",
-            namedImports: [{ name: "CustomLinkProps" }, { name: "CustomLink", importAs: "Link" }],
-          },
-          linkComponent: "Link",
-          linkProps: "CustomLinkProps",
-          hrefProp: "to",
-          generateLinkComponent: true,
-          generateRedirectComponent: true,
-          generateUseRedirect: true,
-          generateUseParams: true,
-          mode: "strict",
-        },
-      });
-
-      expect(templateFile.filename).toBe("LinkLogin");
-      expect(templateFile.extension).toBe(".tsx");
-      expect(templateFile.destinationDir).toBe("path/to/routes");
-      expect(templateFile.template).toMatchInlineSnapshot(`
-        "import React from 'react'
-          import {generateUrl,} from 'route-codegen'
-          import {CustomLinkProps,CustomLink as Link,} from 'src/common/Link'
-          import {patternLogin,UrlParamsLogin,} from './patternLogin'
-          type LinkLoginProps = Omit<CustomLinkProps, 'to'> & { urlParams?: UrlParamsLogin }
-          export const LinkLogin: React.FunctionComponent<LinkLoginProps> = ({ urlParams, ...props }) => {
-            const to = generateUrl(patternLogin, { path: {}, query: urlParams?.query, origin: urlParams?.origin });
-            return <Link {...props} to={to} />;
-          }"
-      `);
-    });
-  });
-});
