--- packages/core/src/generate/generateAppFiles/generatorReactRouterV5/generateRedirectFileReactRouterV5.ts
@@ -1,52 +1,0 @@
-import { TemplateFile, Import } from "../../types";
-import { printImport, capitalizeFirstChar } from "../../utils";
-import { PatternNamedExports } from "../types";
-
-export interface GenerateRedirectFileReactRouterV5Params {
-  routeName: string;
-  destinationDir: string;
-  importGenerateUrl: Import;
-  patternNamedExports: PatternNamedExports;
-}
-
-export const generateRedirectFileReactRouterV5 = (params: GenerateRedirectFileReactRouterV5Params): TemplateFile => {
-  const { routeName: originalRouteName, destinationDir, importGenerateUrl, patternNamedExports } = params;
-
-  const routeName = capitalizeFirstChar(originalRouteName);
-
-  const functionName = `Redirect${routeName}`;
-  const hasPathParams = !!patternNamedExports.pathParamsInterfaceName;
-  const generateUrlFnName = "generateUrl"; // TODO: find a better way to reference this
-
-  const urlParamsModifier = hasPathParams ? "" : "?";
-  const urlParamsTemplate = `urlParams${urlParamsModifier}: ${patternNamedExports.urlParamsInterfaceName}`;
-
-  const template = `${printImport({ defaultImport: "React", from: "react" })}
-  ${printImport(importGenerateUrl)}
-  ${printImport({ namedImports: [{ name: "Redirect" }], from: "react-router" })}
-  ${printImport({
-    namedImports: [{ name: patternNamedExports.urlParamsInterfaceName }, { name: patternNamedExports.patternName }],
-    from: `./${patternNamedExports.filename}`,
-  })}
-  export const ${functionName}: React.FunctionComponent<{ fallback?: React.ReactNode, ${urlParamsTemplate} }> = ({ urlParams, ...props }) => {
-    const to = ${generateUrlFnName}(${patternNamedExports.patternName}, { path: ${
-    hasPathParams ? "urlParams.path" : "{}"
-  }, query: urlParams.query, origin: urlParams.origin });
-    return (
-      <>
-        <Redirect to={to} />
-        {props.fallback}
-      </>
-    );
-  };`;
-
-  return {
-    filename: functionName,
-    destinationDir,
-    extension: ".tsx",
-    template,
-    routeName: originalRouteName,
-    hasDefaultExport: false,
-    hasNamedExports: true,
-  };
-};
