--- packages/core/src/generate/generateAppFiles/generatorReactRouterV5/generatorReactrouterV5.ts
@@ -1,11 +1,0 @@
-import { generateUseParamsFileReactRouterV5 } from "./generateUseParamsFileReactRouterV5";
-import { generateUseRedirectFileReactRouterV5 } from "./generateUseRedirectFileReactRouterV5";
-import { generateLinkFileReactRouterV5 } from "./generateLinkFileReactRouterV5";
-import { generateRedirectFileReactRouterV5 } from "./generateRedirectFileReactRouterV5";
-
-export const generatorReactRouterV5 = {
-  generateUseParamsFile: generateUseParamsFileReactRouterV5,
-  generateUseRedirectFile: generateUseRedirectFileReactRouterV5,
-  generateLinkFile: generateLinkFileReactRouterV5,
-  generateRedirectFile: generateRedirectFileReactRouterV5,
-};
