--- packages/core/src/generate/generateAppFiles/generatorReactRouterV5/index.ts
@@ -1,1 +1,0 @@
-export * from "./generatorReactrouterV5";
