+++ packages/core/src/plugins/typescript-react-router-5/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TypescriptReactRouter5Plugin";
