+++ sample/outputs/default/auth/routes/index.ts
@@ -10,35 +10,35 @@ export * from "./signup/RedirectSignup";
 export * from "./user/patternUser";
 export * from "./user/generateUrlUser";
 export * from "./user/LinkUser";
-export * from "./user/RedirectUser";
 export * from "./user/useRedirectUser";
+export * from "./user/RedirectUser";
 export * from "./account/patternAccount";
 export * from "./account/generateUrlAccount";
 export * from "./account/LinkAccount";
-export * from "./account/RedirectAccount";
 export * from "./account/useRedirectAccount";
+export * from "./account/RedirectAccount";
 export * from "./home/patternHome";
 export * from "./home/generateUrlHome";
 export * from "./home/LinkHome";
-export * from "./home/RedirectHome";
 export * from "./home/useRedirectHome";
+export * from "./home/RedirectHome";
 export * from "./about/patternAbout";
 export * from "./about/generateUrlAbout";
 export * from "./about/LinkAbout";
-export * from "./about/RedirectAbout";
 export * from "./about/useRedirectAbout";
+export * from "./about/RedirectAbout";
 export * from "./contact/patternContact";
 export * from "./contact/generateUrlContact";
 export * from "./contact/LinkContact";
-export * from "./contact/RedirectContact";
 export * from "./contact/useRedirectContact";
+export * from "./contact/RedirectContact";
 export * from "./toc/patternToc";
 export * from "./toc/generateUrlToc";
 export * from "./toc/LinkToc";
-export * from "./toc/RedirectToc";
 export * from "./toc/useRedirectToc";
+export * from "./toc/RedirectToc";
 export * from "./legacy/patternLegacy";
 export * from "./legacy/generateUrlLegacy";
 export * from "./legacy/LinkLegacy";
-export * from "./legacy/RedirectLegacy";
 export * from "./legacy/useRedirectLegacy";
+export * from "./legacy/RedirectLegacy";
