+++ sample/outputs/default/toc/routes/index.ts
@@ -7,40 +7,40 @@ export * from "./toc/useRedirectToc";
 export * from "./user/patternUser";
 export * from "./user/generateUrlUser";
 export * from "./user/LinkUser";
-export * from "./user/RedirectUser";
 export * from "./user/useRedirectUser";
+export * from "./user/RedirectUser";
 export * from "./account/patternAccount";
 export * from "./account/generateUrlAccount";
 export * from "./account/LinkAccount";
-export * from "./account/RedirectAccount";
 export * from "./account/useRedirectAccount";
+export * from "./account/RedirectAccount";
 export * from "./login/patternLogin";
 export * from "./login/generateUrlLogin";
 export * from "./login/LinkLogin";
-export * from "./login/RedirectLogin";
 export * from "./login/useRedirectLogin";
+export * from "./login/RedirectLogin";
 export * from "./signup/patternSignup";
 export * from "./signup/generateUrlSignup";
 export * from "./signup/LinkSignup";
-export * from "./signup/RedirectSignup";
 export * from "./signup/useRedirectSignup";
+export * from "./signup/RedirectSignup";
 export * from "./home/patternHome";
 export * from "./home/generateUrlHome";
 export * from "./home/LinkHome";
-export * from "./home/RedirectHome";
 export * from "./home/useRedirectHome";
+export * from "./home/RedirectHome";
 export * from "./about/patternAbout";
 export * from "./about/generateUrlAbout";
 export * from "./about/LinkAbout";
-export * from "./about/RedirectAbout";
 export * from "./about/useRedirectAbout";
+export * from "./about/RedirectAbout";
 export * from "./contact/patternContact";
 export * from "./contact/generateUrlContact";
 export * from "./contact/LinkContact";
-export * from "./contact/RedirectContact";
 export * from "./contact/useRedirectContact";
+export * from "./contact/RedirectContact";
 export * from "./legacy/patternLegacy";
 export * from "./legacy/generateUrlLegacy";
 export * from "./legacy/LinkLegacy";
-export * from "./legacy/RedirectLegacy";
 export * from "./legacy/useRedirectLegacy";
+export * from "./legacy/RedirectLegacy";
